style: :memo: added eslint perfection to fix the order of imports and vars
