+++ apps/api/src/pkgs/errors/error-schema.ts
@@ -1,4 +1,5 @@
 import { z } from "@hono/zod-openapi";
+
 import { ErrorCode } from "./error-codes";
 
 /**
