+++ apps/api/src/pkgs/hono/types.ts
@@ -1,0 +1,5 @@
+import type { Context as GenericContext } from "hono";
+
+import type { HonoContext } from "./context";
+
+export type Context = GenericContext<HonoContext>;
