+++ apps/api/src/routes/v1-liveness.ts
@@ -1,38 +1,39 @@
 import { createRoute, z } from "@hono/zod-openapi";
+
+import type { App } from "~/pkgs/hono/create-hono-app";
+
 import { openApiErrorResponses } from "~/pkgs/errors/openapi-responses";
 
-import type { App } from "~/pkgs/hono/app";
-
 const route = createRoute({
-  tags: ["liveness"],
-  operationId: "v1.liveness",
   method: "get",
+  operationId: "v1.liveness",
   path: "/v1/liveness",
   responses: {
     200: {
-      description: "The configured services and their status",
       content: {
         "application/json": {
           schema: z.object({
-            status: z.string().openapi({
-              description: "The status of the server",
-            }),
             services: z.object({
-              metrics: z.string().openapi({
-                description: "The name of the connected metrics service",
-                example: "AxiomMetrics",
-              }),
               logger: z.string().openapi({
                 description: "The name of the connected logger service",
                 example: "AxiomLogger or ConsoleLogger",
               }),
+              metrics: z.string().openapi({
+                description: "The name of the connected metrics service",
+                example: "AxiomMetrics",
+              }),
+            }),
+            status: z.string().openapi({
+              description: "The status of the server",
             }),
           }),
         },
       },
+      description: "The configured services and their status",
     },
     ...openApiErrorResponses,
   },
+  tags: ["liveness"],
 });
 
 /**
@@ -56,11 +57,11 @@ export const registerV1Liveness = (app: App) =>
 
     return c.json(
       {
-        status: "we're so back",
         services: {
-          metrics: metrics.constructor.name,
           logger: logger.constructor.name,
+          metrics: metrics.constructor.name,
         },
+        status: "we're so back",
       },
       200,
     );
