+++ apps/api/vitest.integration.ts
@@ -2,19 +2,19 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    include: ["./src/routes/**/*.test.ts", "./src/integration/**/*.test.ts"],
-    reporters: ["html", "verbose"],
-    outputFile: "./.vitest/html",
     alias: {
       "~/": new URL("src/", import.meta.url).pathname,
     },
+    include: ["./src/routes/**/*.test.ts", "./src/integration/**/*.test.ts"],
+    outputFile: "./.vitest/html",
     pool: "threads",
     poolOptions: {
       threads: {
         singleThread: true,
       },
     },
-    testTimeout: 60_000,
+    reporters: ["html", "verbose"],
     teardownTimeout: 60_000,
+    testTimeout: 60_000,
   },
 });
