+++ apps/docs/app/(home)/layout.tsx
@@ -1,6 +1,7 @@
+import type { ReactNode } from "react";
+
 import { baseOptions } from "@/app/layout.config";
 import { HomeLayout } from "fumadocs-ui/layouts/home";
-import type { ReactNode } from "react";
 
 export default function Layout({ children }: { children: ReactNode }) {
   return <HomeLayout {...baseOptions}>{children}</HomeLayout>;
