+++ apps/docs/app/docs/layout.tsx
@@ -1,7 +1,8 @@
+import type { ReactNode } from "react";
+
 import { baseOptions } from "@/app/layout.config";
 import { source } from "@/lib/source";
 import { DocsLayout } from "fumadocs-ui/layouts/docs";
-import type { ReactNode } from "react";
 
 export default function RootDocumentationLayout({
   children,
