+++ apps/docs/app/layout.tsx
@@ -1,7 +1,9 @@
-import { RootProvider } from "fumadocs-ui/provider";
-import "./global.css";
-import { Inter } from "next/font/google";
 import type { ReactNode } from "react";
+
+import "./global.css";
+
+import { RootProvider } from "fumadocs-ui/provider";
+import { Inter } from "next/font/google";
 
 const inter = Inter({
   subsets: ["latin"],
