+++ apps/docs/postcss.config.mjs
@@ -1,6 +1,6 @@
 export default {
   plugins: {
-    tailwindcss: {},
     autoprefixer: {},
+    tailwindcss: {},
   },
 };
