+++ eslint.config.mjs
@@ -1,5 +1,6 @@
 import pluginJs from "@eslint/js";
 import biome from "eslint-config-biome";
+import perfectionist from "eslint-plugin-perfectionist";
 // import pluginReact from "eslint-plugin-react";
 // import tailwind from "eslint-plugin-tailwindcss";
 import eslintPluginUnicorn from "eslint-plugin-unicorn";
@@ -20,7 +21,9 @@ export default [
     ],
   },
   { files: ["**/*.{js,mjs,cjs,ts,jsx,tsx}"] },
-  { languageOptions: { globals: globals.browser } },
+  {
+    languageOptions: { globals: globals.browser },
+  },
   pluginJs.configs.recommended,
   eslintPluginUnicorn.configs["flat/recommended"],
   ...tseslint.configs.recommended,
@@ -36,12 +39,58 @@ export default [
         "error",
         {
           allowList: {
+            generateStaticParams: true,
             getInitialProps: true,
             getStaticProps: true,
-            generateStaticParams: true,
           },
         },
       ],
+    },
+  },
+  {
+    plugins: {
+      perfectionist,
+    },
+    rules: {
+      "perfectionist/sort-imports": [
+        "error",
+        {
+          customGroups: { type: {}, value: {} },
+          environment: "node",
+          groups: [
+            "type",
+            ["builtin", "external"],
+            "internal-type",
+            "internal",
+            ["parent-type", "sibling-type", "index-type"],
+            ["parent", "sibling", "index"],
+            "object",
+            "unknown",
+          ],
+          ignoreCase: true,
+          internalPattern: ["^~/.+"],
+          maxLineLength: undefined,
+          newlinesBetween: "always",
+          order: "asc",
+          partitionByComment: false,
+          partitionByNewLine: false,
+          specialCharacters: "keep",
+          type: "alphabetical",
+        },
+      ],
+      "perfectionist/sort-interfaces": ["error"],
+      "perfectionist/sort-objects": [
+        "error",
+        {
+          type: "alphabetical",
+        },
+      ],
+    },
+    settings: {
+      perfectionist: {
+        partitionByComment: true,
+        type: "ne",
+      },
     },
   },
   biome,
