+++ internals/logs/src/types.ts
@@ -1,4 +1,5 @@
 import type { z } from "zod";
+
 import type { logSchema } from "./logs";
 
 /**
