+++ internals/metrics/src/metrics.ts
@@ -6,50 +6,50 @@ import { z } from "zod";
  */
 export const metricSchema = z.discriminatedUnion("metric", [
   z.object({
-    metric: z.literal("metric.http.request"),
+    city: z.string().optional(),
+    coldStart: z.boolean().optional(),
+    error: z.string().optional(),
     host: z.string(),
-    path: z.string(),
+    isolateId: z.string().optional(),
     method: z.string(),
+    metric: z.literal("metric.http.request"),
+    path: z.string(),
     status: z.number(),
+    /**
+     * Regional data might differ on non-Cloudflare deployments.
+     * @type {string | undefined}
+     */
+    colo: z.string().optional(),
+    context: z.record(z.unknown()),
+    continent: z.string().optional(),
+    country: z.string().optional(),
+    fromAgent: z.string().optional(),
     /**
      * Milliseconds since the worker was initialized for the first time.
      * A non-zero value indicates the worker is reused.
      * @type {number | undefined}
      */
     isolateLifetime: z.number().optional(),
-    isolateId: z.string().optional(),
-    error: z.string().optional(),
-    coldStart: z.boolean().optional(),
     serviceLatency: z.number(),
-    /**
-     * Regional data might differ on non-Cloudflare deployments.
-     * @type {string | undefined}
-     */
-    colo: z.string().optional(),
-    continent: z.string().optional(),
-    country: z.string().optional(),
-    city: z.string().optional(),
     userAgent: z.string().optional(),
-    fromAgent: z.string().optional(),
-    context: z.record(z.unknown()),
   }),
   z.object({
+    dbRes: z.string().optional(),
+    latency: z.number(),
     metric: z.literal("metric.db.read"),
     query: z.enum([
       "getKeyAndApiByHash",
       "loadFromOrigin",
       "getKeysByKeyAuthId",
     ]),
-    latency: z.number(),
-    dbRes: z.string().optional(),
     sql: z.string().optional(),
   }),
   z.object({
+    attempts: z.number().optional(),
+    latency: z.number(),
     metric: z.literal("metric.db.transaction"),
     name: z.string(),
     path: z.string().optional(),
-    latency: z.number(),
-    attempts: z.number().optional(),
   }),
 ]);
