+++ package.json
@@ -18,6 +18,7 @@
     "@eslint/js": "^9.15.0",
     "eslint": "^9.15.0",
     "eslint-config-biome": "^1.9.3",
+    "eslint-plugin-perfectionist": "^4.1.2",
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-tailwindcss": "^3.17.5",
     "eslint-plugin-unicorn": "^56.0.1",
